Splitted packedbubble series into several files.
