+++ ts/Series/PackedBubble/PackedBubblePointOptions.d.ts
@@ -1,0 +1,29 @@
+/* *
+ *
+ *  (c) 2010-2018 Grzegorz Blachlinski, Sebastian Bochan
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type BubblePointOptions from '../Bubble/BubblePointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface PackedBubblePointOptions extends BubblePointOptions {
+    mass?: number;
+}
+
+export default PackedBubblePointOptions;
